Rename and reorganize Google News and test files to examples directory
